+++ tick/stateful/eval_binary_node.go
@@ -154,6 +154,10 @@ func (n *EvalBinaryNode) EvalTime(scope *Scope, executionState ExecutionState) (
 	return time.Time{}, ErrTypeGuardFailed{RequestedType: ast.TTime, ActualType: n.constReturnType}
 }
 
+func (n *EvalBinaryNode) EvalMissing(scope *Scope, executionState ExecutionState) (*ast.Missing, error) {
+	return nil, ErrTypeGuardFailed{RequestedType: ast.TMissing, ActualType: n.constReturnType}
+}
+
 func (e *EvalBinaryNode) EvalDuration(scope *Scope, executionState ExecutionState) (time.Duration, error) {
 	result, err := e.eval(scope, executionState)
 	if err != nil {
@@ -182,7 +186,13 @@ func (e *EvalBinaryNode) EvalString(scope *Scope, executionState ExecutionState)
 
 // EvalBool executes the expression based on eval bool
 func (e *EvalBinaryNode) EvalBool(scope *Scope, executionState ExecutionState) (bool, error) {
-	result, err := e.eval(scope, executionState)
+	var result resultContainer
+	var err *ErrSide
+	if e.leftEvaluator.IsDynamic() || e.rightEvaluator.IsDynamic() {
+		result, err = e.evaluateDynamicNode(scope, executionState, e.leftEvaluator, e.rightEvaluator)
+	} else {
+		result, err = e.eval(scope, executionState)
+	}
 	if err != nil {
 		return false, err.error
 	}
@@ -280,12 +290,13 @@ func (e *EvalBinaryNode) evaluateDynamicNode(scope *Scope, executionState Execut
 	// For example: "count() == 1"
 	//  1. we evaluate the left side and counter is 1 (upper ^ in this function)
 	//  2. we evaluate the second time in "EvalBool"
+	typeExecutionState := CreateExecutionState()
 
-	if leftType, err = left.Type(scope, executionState); err != nil {
+	if leftType, err = left.Type(scope, typeExecutionState); err != nil {
 		return emptyResultContainer, &ErrSide{error: err, IsLeft: true}
 	}
 
-	if rightType, err = right.Type(scope, executionState); err != nil {
+	if rightType, err = right.Type(scope, typeExecutionState); err != nil {
 		return emptyResultContainer, &ErrSide{error: err, IsRight: true}
 	}